backend/requirements.txt
alembic==1.8.1
anyio==3.6.2
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
log-utils==0.4.1
Mako==1.2.4
MarkupSafe==2.1.1
pika==1.3.1
pika-stubs==0.1.3
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tornado==6.2
typing_extensions==4.4.0
uvicorn==0.20.0
